add md5 and umbers section
