+++ _content/tour/algorithms-md5.article
@@ -2,3 +2,5 @@ MD5
 TODO - Description
 
 * MD5
+
+.play algorithms/md5/dlhash.go
