+++ _content/tour/algorithms-numbers.article
@@ -2,3 +2,7 @@ Numbers
 TODO - Description
 
 * Numbers
+
+** Palindrome
+
+.play algorithms/numbers/palindrome/palindrome.go
